+++ tests/common/testfs/mod.rs
@@ -1,3 +1,3 @@
 #[macro_use]
 pub mod fuse;
-pub mod utils;
+pub mod util;
